refactor: 음식점 추가 및 모달 기능 개선 - 이벤트 핸들러 통합 및 button UI와 도메인 분리
