+++ src/components/form/createRestaurantForm.js
@@ -2,17 +2,65 @@ import $button from "../common/button.js";
 import $buttonContainer from "../layout/button-container.js";
 import $inputItem from "./input-item.js";
 import $form from "./form.js";
-import { addRestaurant } from "./formEvent.js";
+import $restaurantItem from "../restaurant/restaurant-item.js";
 import { UI_CONFIG } from "../../constants/uiConfig.js";
 import { FORM_FIELDS } from "../../constants/formFields.js";
-import { handleModalOpen } from "../modal/modal.js";
+import { handleModalClose, handleModalOpen } from "../modal/modal.js";
+import { validateRestaurantForm } from "../../validation/validationForm.js";
+import { storageHandler } from "../../utils/storageHandler.js";
+import { STORAGE_KEY_NAME } from "../../constants/storage.js";
+import { CATEGORY_ICON } from "../../constants/iconPath.js";
+
+const restaurantFormReset = () => {
+  handleModalClose();
+  const form = document.getElementById("add-restaurant-form");
+  form.reset();
+};
+
+const addRestaurant = (data) => {
+  const newRestaurant = {
+    categoryIcon: CATEGORY_ICON[data.category],
+    categoryTitle: data.category,
+    name: data.name,
+    distance: `캠퍼스부터 ${data.distance}분 내`,
+    description: data.description,
+    link: data.link,
+    id: new Date(),
+  };
+  document
+    .querySelector(".restaurant-list")
+    .prepend($restaurantItem(newRestaurant));
+
+  const currentItem = storageHandler.getItem(STORAGE_KEY_NAME);
+  storageHandler.setItem(STORAGE_KEY_NAME, [...currentItem, newRestaurant]);
+
+  const noRestaurant = document.getElementById("noRestaurant");
+  if (noRestaurant) noRestaurant.remove();
+};
+
+const handleAddRestaurant = (e) => {
+  e.preventDefault();
+
+  try {
+    const form = document.getElementById("add-restaurant-form");
+    const formData = new FormData(form);
+    const data = Object.fromEntries(formData);
+    validateRestaurantForm(form);
+    addRestaurant(data);
+    restaurantFormReset(form);
+  } catch (error) {
+    alert(error.message);
+    console.log(error);
+  }
+};
 
 const $createRestaurantForm = () => {
   const container = document.querySelector(".modal-container");
   container.replaceChildren();
 
+  const cancelEvent = { eventType: "click", eventHandler: restaurantFormReset };
   const submitCancelButtons = $buttonContainer([
-    $button(UI_CONFIG.BUTTONS.CANCEL),
+    $button(UI_CONFIG.BUTTONS.CANCEL, cancelEvent),
     $button(UI_CONFIG.BUTTONS.ADD),
   ]);
 
@@ -30,7 +78,8 @@ const $createRestaurantForm = () => {
   title.textContent = "새로운 음식점";
 
   container.appendChild(title);
-  container.appendChild($form(restaurantAddForm, addRestaurant));
+  const submitForm = { eventType: "submit", eventHandler: handleAddRestaurant };
+  container.appendChild($form(restaurantAddForm, submitForm));
 
   handleModalOpen();
 };
