+++ src/constants/uiConfig.js
@@ -1,29 +1,4 @@
-// import { handleAddRestaurant } from "../components/form/formEvent.js";
-import { addRestaurant } from "../components/form/formEvent.js";
-import { handleModalClose } from "../components/modal/modal.js";
 import { deepFreeze } from "../utils/deepFreeze.js";
-import { validateRestaurantForm } from "../validation/validationForm.js";
-
-const restaurantFormReset = () => {
-  handleModalClose();
-  const form = document.getElementById("add-restaurant-form");
-  form.reset();
-};
-
-const handleAddRestaurant = (e) => {
-  e.preventDefault();
-
-  try {
-    const form = document.getElementById("add-restaurant-form");
-    const formData = new FormData(form);
-    const data = Object.fromEntries(formData);
-    validateRestaurantForm(form);
-    addRestaurant(data);
-    restaurantFormReset(form);
-  } catch (error) {
-    alert(error.message);
-  }
-};
 
 export const UI_CONFIG = deepFreeze({
   HEADER: {
@@ -34,8 +9,6 @@ export const UI_CONFIG = deepFreeze({
   BUTTONS: {
     CANCEL: {
       text: "취소하기",
-      eventType: "click",
-      event: restaurantFormReset,
       attribute: {
         type: "button",
         className: "button button--secondary text-caption cancel-button",
@@ -43,7 +16,6 @@ export const UI_CONFIG = deepFreeze({
     },
     ADD: {
       text: "추가하기",
-      event: handleAddRestaurant,
       attribute: {
         id: "addRestaurantButton",
         type: "submit",
@@ -59,14 +31,11 @@ export const UI_CONFIG = deepFreeze({
       attribute: {
         id: "deleteRestaurantButton",
         type: "button",
-        disabled: true,
         className: "button button--secondary text-caption cancel-button",
       },
     },
     CLOSE: {
       text: "닫기",
-      eventType: "click",
-      event: handleModalClose,
       attribute: {
         id: "closeModalButton",
         type: "button",
