+++ src/utils/storageHandler.js
@@ -1,4 +1,5 @@
 export const storageHandler = {
   getItem: (data) => JSON.parse(localStorage.getItem(data) ?? "[]") || [],
   setItem: (key, value) => localStorage.setItem(key, JSON.stringify(value)),
+  deleteItem: (key, value) => localStorage.setItem(key, JSON.stringify(value)),
 };
